feat: automatic table migrations
Adding a mechanism to create tables before they're
needed and migrate the DynamoDB. All scripts
that need persistent storage should add to the
migration file before shipping out.
