+++ scripts/karma/index.ts
@@ -1,6 +1,5 @@
 import type AppCustomContext from "@slack/bolt/dist/App.d.ts";
-//import { DynamoDBClient, UpdateItemCommand, PutItemCommand } from '@aws-sdk/client-dynamodb';
-import { deltaReputation, writeReputationToDynamoDB } from "./countEntries";
+import { deltaReputation, writeReputationToDynamoDB } from "./reputation";
 import { generateKarmaMessages } from "./announce";
 /**
  * This function controls the enabled
@@ -8,7 +7,6 @@ import { generateKarmaMessages } from "./announce";
  * he is running (locally or in a lambda)
  */
 const karma = async (app: AppCustomContext) => {
-  //const dynamoDbClient = new DynamoDBClient({ region: 'us-east-1' });
   // Listens to incoming messages that contain "hello"
   app.message(/([\w<@>]+)\s*([\+\-]{2})/g, async ({ message, say }) => {
     if (message.subtype === undefined || message.subtype === "bot_message") {
